Remove Android-specific annotation from the JVM artifact (#43)
The project is a bit stale, AndroidX came out a long time ago and we are still on the support-level dependencies. However, instead of using AndroidX I suggest to remove support dependencies altogether. This will be better from the maintenance POV.

This change removes an Android-specific annotation from a JVM (Java) artifact. Also the annotation was set on an internal class nobody should use in the first place. 

PTAL @ameliariely, @amphora001, @artem-zinnatullin